# contracts/openapi-v0.6.yaml
openapi: 3.1.0
info: {title: PrivatePhuket API, version: 0.6.0}
servers: [{url: /api/v1}]
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
security: [{bearerAuth: []}]
paths:
...
  /properties:
    get: {responses: {'200': {description: Property list}}}
    post: {responses: {'201': {description: Created}, '403': {description: Forbidden}}}
  /properties/{id}/workflow:
    post: {parameters: [{name: id,in: path,required: true,schema: {type: string,format: uuid}}],responses: {'200': {description: Transitioned},'409': {description: Invalid transition}}}
...
  /properties/{id}/calculate:
    post: {responses: {'201': {description: Immutable calculation run},'422': {description: Blocked}}}
...
  /properties/{id}/publish:
    post: {responses: {'201': {description: Published},'409': {description: Run not FINAL or APPROVED}}}
...